extract each type of product into its own file
